+++ src/pageLayouts/404/index.js
@@ -1,3 +1,3 @@
-import NotFoundPage from './404Page.js';
+import NotFoundPage from './404.js';
 
 export default NotFoundPage;
